+++ src/util.js
@@ -113,6 +113,7 @@ function _global_options(argv) {
   let flags = [];
   /* Boolean type */
   flags.push(def_flag(argv.global, '-g'));
+  flags.push(def_flag(argv.config, '-c'));
   flags.push(def_flag(argv.all, '-a'));
   flags.push(def_flag(argv.quick, '-q'));
   flags.push(def_flag(argv.force, '-f'));
@@ -147,6 +148,20 @@ function el_script(name) {
 }
 
 /**
+ * Get the working environment name.
+ * @param { JSON } argv - Argument vector.
+ * @return Return a string represent the current working environment.
+ */
+function _environment_name (argv) {
+  if (argv.global)
+    return 'global (~/.eask/)';
+  else if (argv.config)
+    return 'configuration (~/.emacs.d/)';
+  else
+    return 'development (./)';
+}
+
+/**
  * Call emacs process
  * @param { string } script - name of the script from `../lisp`
  * @param { string } args - the rest of the arguments
@@ -161,7 +176,7 @@ async function e_call(argv, script, ...args) {
     let cmd = cmd_base.concat(cmd_args).concat(cmd_global);
     cmd = _remove_undefined(cmd);
 
-    let env_status = (argv.global) ? 'global' : 'development';
+    let env_status = _environment_name(argv);
     console.log(`Running Eask in the ${env_status} environment`);
     console.log('Press Ctrl+C to cancel.');
     console.log('');