+++ test/commands/config/run.sh
@@ -1,3 +1,5 @@
+#!/usr/bin/env bash
+
 # Copyright (C) 2022-2023 Jen-Chieh Shen
 
 # This program is free software; you can redistribute it and/or modify
@@ -17,9 +19,22 @@
 
 ## Commentary:
 #
-# Copy configuration over to ~/.emacs.d/
+# Here we test all config (~/.emacs.d/) that the Emacser can be use daily!
+#
+# Notice, to make config commands work; we need a minimum configuration
+# (mini.emacs.d), and place it under to the default Emacs directory!
 #
 
+set -e
+
 echo "Copy test configuration"
-mkdir "$env:USERPROFILE/AppData/Roaming/.emacs.d"
-robocopy /e "./test/fixtures/mini.emacs.d/" "$env:USERPROFILE/AppData/Roaming/.emacs.d"
+./test/fixtures/mini.emacs.d/scripts/copy_config.sh
+
+echo "Testing config (~/.emacs.d/) commands..."
+eask archives  -c
+
+eask install   -c spinner ivy beacon company fuzzy
+eask uninstall -c ivy fuzzy
+
+eask list      -c --depth=0
+eask outdated  -c