+++ test/commands/global/run.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# Copyright (C) 2022-2023 Jen-Chieh Shen
+# Copyright (C) 2023 Jen-Chieh Shen
 
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -19,18 +19,18 @@
 
 ## Commentary:
 #
-# Here we test all global (~/.emacs.d/) that the Emacser can be use daily!
+# Here we test all global (~/.eask/) that the Emacser can be use daily!
 #
-# Notice, to make global commands work; we need a minimum configuration
+# Notice, to make config commands work; we need a minimum configuration
 # (mini.emacs.d), and place it under to the default Emacs directory!
 #
 
 set -e
 
 echo "Copy test configuration"
-./test/fixtures/mini.emacs.d/scripts/copy_config.sh
+./test/fixtures/mini.eask/scripts/copy_eask.sh
 
-echo "Testing global commands..."
+echo "Testing global (~/.eask/) commands..."
 eask archives  -g
 
 eask install   -g spinner ivy beacon company fuzzy